backend/micro-services/business-service/src/main/java/com/pocketpay/businessservice/controller/SubCategoryController.java
package com.pocketpay.businessservice.controller;

import com.pocketpay.businessservice.dto.SubCategoryDto;
import com.pocketpay.businessservice.service.SubCategoryService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;

@RestController
@RequestMapping("/subCategories")
public class SubCategoryController {
    @Autowired
    private SubCategoryService subcategoryService;

    @GetMapping
    public ResponseEntity<List<SubCategoryDto>> getAllSubCategories() {
        List<SubCategoryDto> subcategories = subcategoryService.getAllSubCategories();
        return new ResponseEntity<>(subcategories, HttpStatus.OK);
    }

    @GetMapping("/{id}")
    public ResponseEntity<SubCategoryDto> getSubCategoryById(@PathVariable int id) {
        SubCategoryDto subcategory = subcategoryService.getSubCategoryById(id);
        if (subcategory != null) {
            return new ResponseEntity<>(subcategory, HttpStatus.OK);
        } else {
            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
        }
    }
}
